Fix CI tests
`cljstyle fix` and remove cloverage.

The cloverage build always fails, I think because the project has no
tests, so I'm removing it.

I had to twiddle some bits in CircleCI to get this building for some
reason, so I didn't test this before submitting. (I thought CircleCI not
running was due to the GitHub service degredation this morning but that
was not the case.)